requirements.txt
tinify
boto3
tqdm
python-dotenv